% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to your plot or image
This function places a logo at one of the corners of your image/plot}
\usage{
add_logo(
  plot_path,
  logo_path = c("brookings", "es", "gs", "fp", "metro", "global", "bi", "bc", "cc",
    "ccf", "ceaps", "cepm", "chp", "cmep", "csd", "cti", "crm", "cue", "cuse", "doha",
    "hc", "tsinghua"),
  logo_position = "bottom left",
  logo_scale = 5,
  height_padding = 0.045,
  width_padding = 0.01
)
}
\arguments{
\item{plot_path}{url or local file for the plot}

\item{logo_path}{url, local file for the logo, or one of the built in logos. Can be a character string in c('brookings','es', 'gs', 'fp', 'metro', 'global', 'bi', 'bc', 'cc', 'ccf', 'ceaps', 'cepm', 'chp', 'cmep', 'csd', 'cti', 'crm', 'cue', 'cuse', 'doha', 'hc', 'tsinghua')}

\item{logo_position}{Optional. Valid options are 'top left', 'top right', 'bottom left', and 'bottom right'. Default logo position is 'bottom right'.}

\item{logo_scale}{Optional. The portion of the plot width the logo should occupy. The default is 5 which means that the logo will be 1/5th of the plot width.}

\item{height_padding}{Optional. Control the y axis position of a logo by specifying the padding proportion. Default is 4.5\% padding.}

\item{width_padding}{Optional. Control the x axis position of a logo by specifying the padding proportion. Default is 1\% padding.}
}
\value{

}
\description{
Add a logo to your plot or image
This function places a logo at one of the corners of your image/plot
}
