% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{brookings_cols}
\alias{brookings_cols}
\title{Function to extract brookings colors as hex codes}
\usage{
brookings_cols(...)
}
\arguments{
\item{...}{Character names of Brookings_colors}
}
\description{
Function to extract brookings colors as hex codes
}
