% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{brookings_pal}
\alias{brookings_pal}
\title{Return function to interpolate a brookings color palette}
\usage{
brookings_pal(palette = "brand1", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in brookings_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a brookings color palette
}
