% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{brookings_colors}
\alias{brookings_colors}
\alias{brookings_palettes}
\title{Brookings Institution palettes}
\format{
An object of class \code{character} of length 94.

An object of class \code{list} of length 26.
}
\source{
Brookings Style Guide
x
}
\usage{
brookings_colors

brookings_palettes
}
\description{
Vectors with hex-color codes that correspond to the color palettes outlined in the Brookings Style Guide.
x
}
\concept{brookings palettes}
\keyword{datasets}
