% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brookings_save.R
\name{brookings_save}
\alias{brookings_save}
\title{Save ggplot2 plots in standard Brookings sizes}
\usage{
brookings_save(
  filename,
  size = "medium",
  dpi = 300,
  height = NULL,
  plot = ggplot2::last_plot(),
  ...
)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{size}{Size of plot. Must be one of "small" (3.25 x 2 inches),
"medium" (6.5 x 4 inches), or "large" (9 x 6.5 inches).}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{height}{Adjust height manually}

\item{plot}{Plot to save, defaults to last plot displayed.}

\item{...}{Pass additional arguments to ggplot2::ggsave()}
}
\description{
This is a function to save standardized images for Brookings publications.
For more functionality like custom widths and heights, use ggsave() directly.
}
\examples{

\dontrun{
library(tidyverse)
library(ggbrookings)

theme_set(theme_brookings())

cars \%>\%
  ggplot(aes(speed, dist)) +
  geom_point()

brookings_save("test.png")
}

}
