% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_palette.R
\name{brookings_view_palette}
\alias{brookings_view_palette}
\title{View brookings palettes}
\usage{
brookings_view_palette(palette = "brand1")
}
\arguments{
\item{palette}{Options are \code{brand1}, \code{brand2}, \code{analogous1}, \code{analogous2}, \code{contrasting1}, \code{contrasting2}, \code{gender1}, \code{gender2}, \code{gender3}, \code{pos_neg1}, \code{pos_neg2}, \code{political1}, \code{political2}, \code{political3}, \code{political4}, \code{categorical}, \code{sequential1}, \code{sequential2}, \code{diverging}, and \code{misc}}
}
\description{
\code{brookings_view_palette} displays the colors and hexadecimal codes for colors in brookings palettes.
}
\section{Palettes}{

\describe{
\item{Analogous}{Different shades of the same hue, or of similar hues can be used when the associated values are related.}
\item{Contrasting}{Colors on the opposite ends of the spectrum. Use Brookings Blue with Secondary colors.}
\item{Semantic}{Where applicable, use colors that are associated with certain concepts. For e.g., semantic1, semantic2, and semantic3 could show subsets of gender data (female, male and other).}
\item{Positive & Negative}{Shows pros, cons and neutral, or positive, negative and neutral data.}
\item{Political}{Use red and blue of similar intensity to represent data related to political parties in the US. Yellow in political3 and political4 represents ‘Independent’ category}
\item{Misc}{A pleasing option using Brookings Blue and accent yellow.}
}
}

\examples{
brookings_view_palette()
brookings_view_palette("brand1")
}
