% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_inter}
\alias{import_inter}
\title{Import and register Inter font}
\usage{
import_inter()
}
\description{
\code{import_inter()} tests to see if Inter is imported and registered. If
Inter, isn't imported and registered, then \code{import_inter()} imports and
registers Inter with
\href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Inter must be installed on your computer for \code{import_inter()} to
work. Inter is Brooking's main font. To install, visit
\href{https://fonts.google.com/specimen/Inter}{Google fonts} and click
"Download family". Unzip and open each of the .ttf files and click install.
}
