% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_roboto}
\alias{import_roboto}
\title{Import and register Roboto font}
\usage{
import_roboto()
}
\description{
\code{import_roboto()} tests to see if Roboto is imported and registered. If
Roboto, isn't imported and registered, then \code{import_roboto()} imports and
registers Roboto with
\href{https://github.com/wch/extrafont}{library(extrafont)}.
}
\details{
Note: Roboto must be installed on your computer for \code{import_roboto()} to
work. Roboto is Brooking's main font. To install, visit
\href{https://fonts.google.com/specimen/roboto}{Google fonts} and click
"Download family". Unzip and open each of the .ttf files and click install.

Test to see if roboto is imported and registered with \code{roboto_test()}.
}
