% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{import_thp_fonts}
\alias{import_thp_fonts}
\title{Import THP fonts}
\usage{
import_thp_fonts()
}
\description{
Imports Photina, Helvetica Neue, and Myriad Pro
}
\details{
import_*() functions taken from hrbrthemes.
}
\seealso{
\code{\link[extrafont]{font_import}}
}
