% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{register_thp_fonts}
\alias{register_thp_fonts}
\title{Register Hamilton Fonts}
\usage{
register_thp_fonts(dpi = 300)
}
\arguments{
\item{dpi}{set dpi (default is 300)}
}
\description{
Register Hamilton Fonts
}
