% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_color_brookings}
\alias{scale_color_brookings}
\title{Color scale constructor for brookings colors}
\usage{
scale_color_brookings(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of brookings_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(),
used respectively when discrete is TRUE or FALSE}
}
\value{

}
\description{
Color scale constructor for brookings colors
}
