% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_brookings}
\alias{theme_brookings}
\title{Title}
\usage{
theme_brookings(
  base_size = 14,
  base_family = "Roboto",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  ticks = TRUE
)
}
\value{

}
\description{
Title
}
