% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_thp}
\alias{theme_thp}
\title{Title}
\usage{
theme_thp(
  base_size = 14,
  base_family = "PhotinaMTW04-Regular",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
