% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to your plot or image
This function places a logo at one of the corners of your image/plot}
\usage{
add_logo(
  plot_path,
  logo_path = c("brookings", "es", "gs", "fp", "metro", "global", "bi", "bc", "cc",
    "ccf", "ceaps", "cepm", "chp", "cmep", "csd", "cti", "crm", "cue", "cuse", "doha",
    "hc", "tsinghua"),
  logo_position = "bottom left",
  logo_scale = 5,
  height_padding = 0.045,
  width_padding = 0.01
)
}
\arguments{
\item{plot_path}{url or local file for the plot}

\item{logo_path}{url, local file for the logo, or one of the built in logos.}

\item{width_padding}{}
}
\value{

}
\description{
Add a logo to your plot or image
This function places a logo at one of the corners of your image/plot
}
