% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbrookings-package.R
\docType{package}
\name{ggbrookings-package}
\alias{ggbrookings}
\alias{ggbrookings-package}
\title{ggbrookings: Grammar of Graphics in Brookings Style}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

An implementation of the Brookings style guide using ggplot2.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://brookingsinstitution.github.io/ggbrookings/}
  \item \url{https://github.com/BrookingsInstitution/ggbrookings}
  \item Report bugs at \url{https://github.com/BrookingsInstitution/ggbrookings/issues}
}

}
\author{
\strong{Maintainer}: Valerie Wirtschafter \email{vwirtschafter@brookings.edu} (\href{https://orcid.org/ANOTHER-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Manuel Alcalá Kovalski \email{malcalakovalski@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
\keyword{internal}
