% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scales_brookings}
\alias{scales_brookings}
\alias{scale_color_brookings}
\alias{scale_fill_brookings}
\title{Color scales in the Brookings style}
\usage{
scale_color_brookings(
  palette = "brand1",
  discrete = TRUE,
  reverse = FALSE,
  ...
)

scale_fill_brookings(palette = "brand1", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of brookings_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
Color scales in the Brookings style
}
\section{Palettes}{

\describe{
\item{Analogous}{Different shades of the same hue, or of similar hues can be used when the associated values are related.}
\item{Contrasting}{Colors on the opposite ends of the spectrum. Use Brookings Blue with Secondary colors.}
\item{Semantic}{Where applicable, use colors that are associated with certain concepts. For e.g., semantic1, semantic2, and semantic3 could show subsets of gender data (female, male and other)}.
\item{Positive & Negative}{Shows pros, cons and neutral, or positive, negative and neutral data.}
\item{Political}{Use red and blue of similar intensity to represent data related to political parties in the US. Yellow in political3 and political4 represents ‘Independent’ category}
\item{Misc}{A pleasing option using Brookings Blue and accent yellow.}
}
}

