% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_thp}
\alias{theme_thp}
\title{Title}
\usage{
theme_thp(
  base_size = 12,
  base_family = "Calibri",
  base_line_size = base_size/24,
  base_rect_size = base_size/24,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family. Default is "Inter".}

\item{base_line_size}{line size in mm.}

\item{base_rect_size}{border size in mm.}

\item{...}{additional arguments passed to theme_brookings()}
}
\description{
Title
}
